Crate bonsaidb_keystorage_s3

VaultKeyStorage using S3-compatible storage.

This is the recommended method for securing your BonsaiDb database. There are many ways to acquire secure, inexpensive S3-compatible storage, such as Backblaze B2.

Do not configure your bucket with public access. You should only allow access from the IP addresses that your BonsaiDb server(s) are hosted on, or only allow authenticated access.

To use this, specify the vault_key_storage configuration parameter:

vault_key_storage: Some(Box::new(S3VaultKeyStorage::from(
        s3::Bucket::new(
            "bucket-name",
            s3::Region::Custom {
                endpoint: String::from("s3.us-west-001.backblazeb2.com"),
                region: String::default(),
            },
            s3::creds::Credentials::new(Some(&"access key id"), Some(&"secret access key"), None, None, None).unwrap(),
        )
        .unwrap(),
    ))),

The API calls are performed by the [s3] crate.

Re-exports

pub use s3;

Structs

S3VaultKeyStorage

S3-compatible VaultKeyStorage implementor.