Crate bonsaidb_keystorage_s3

[VaultKeyStorage] using S3-compatible storage.

This is the recommended method for securing your BonsaiDb database. There are many ways to acquire secure, inexpensive S3-compatible storage, such as Backblaze B2.

Do not configure your bucket with public access. You should only allow access from the IP addresses that your BonsaiDb server(s) are hosted on, or only allow authenticated access.

To use this, specify the vault_key_storage configuration parameter:

let directory = TestDirectory::new("bonsaidb-keystorage-s3-basic");
let configuration = StorageConfiguration::new(&directory)
    .vault_key_storage(
        S3VaultKeyStorage::new("bucket_name").endpoint(Endpoint::immutable(
            Uri::try_from("https://s3.us-west-001.backblazeb2.com").unwrap(),
        )),
    )
    .default_encryption_key(KeyId::Master);

The API calls are performed by the aws-sdk-s3 crate.

Re-exports

pub use aws_sdk_s3;

pub use http;

Structs

S3VaultKeyStorage

S3-compatible [VaultKeyStorage] implementor.